move NonEmbeddedHint to subfolder
