+++ test/modules/iframe/components/NonEmbeddedHint.test.js
@@ -1,5 +1,5 @@
 /* eslint-disable no-undef */
-import { NonEmbeddedHint } from '../../../../src/modules/iframe/components/NonEmbeddedHint';
+import { NonEmbeddedHint } from '../../../../src/modules/iframe/components/hint/NonEmbeddedHint';
 import { TestUtils } from '../../../test-utils';
 import { $injector } from '../../../../src/injection';
